+++ src/rpc/cfx.js
@@ -49,17 +49,17 @@ class CFX extends RPCMethodFactory {
       },
       {
         method: 'cfx_maxPriorityFeePerGas',
-        alias: 'maxPriorityFeePerGas',
         responseFormatter: format.bigUInt,
       },
       {
         method: 'cfx_getFeeBurnt',
-        alias: 'getFeeBurnt',
+        requestFormatters: [
+          format.epochNumberOrUndefined,
+        ],
         responseFormatter: format.bigUInt,
       },
       {
         method: 'cfx_feeHistory',
-        alias: 'feeHistory',
         requestFormatters: [
           format.bigUIntHex,
           format.epochNumber,
@@ -446,8 +446,6 @@ class CFX extends RPCMethodFactory {
   async populateTransaction(options) {
     const {
       defaultGasPrice,
-      defaultGasRatio,
-      defaultStorageRatio,
     } = this.conflux;
 
     options.from = this._formatAddress(options.from);
@@ -469,6 +467,10 @@ class CFX extends RPCMethodFactory {
       options.epochHeight = await this.epochNumber();
     }
 
+    if (options.gasPrice && (options.maxFeePerGas || options.maxPriorityFeePerGas)) {
+      throw new Error('`gasPrice` should not be set with `maxFeePerGas` or `maxPriorityFeePerGas`');
+    }
+
     // auto detect transaction type
     let baseFeePerGas;
     if (options.type === undefined) {
@@ -488,16 +490,9 @@ class CFX extends RPCMethodFactory {
         gas = CONST.TRANSACTION_GAS;
         storageLimit = CONST.TRANSACTION_STORAGE_LIMIT;
       } else {
-        const { gasUsed, storageCollateralized, gasLimit } = await this.estimateGasAndCollateral(options);
-        if (defaultGasRatio) {
-          gas = format.big(gasUsed).times(defaultGasRatio).toFixed(0);
-          if (gas > 15000000) {
-            gas = 15000000;
-          }
-        } else {
-          gas = gasLimit;
-        }
-        storageLimit = format.big(storageCollateralized).times(defaultStorageRatio).toFixed(0);
+        const { gasUsed, storageCollateralized } = await this.estimateGasAndCollateral(options);
+        gas = gasUsed;
+        storageLimit = storageCollateralized;
       }
 
       if (options.gas === undefined) {
@@ -519,18 +514,32 @@ class CFX extends RPCMethodFactory {
           options.gasPrice = defaultGasPrice;
         }
       }
+      options.maxFeePerGas = undefined;
+      options.maxPriorityFeePerGas = undefined;
     }
     // auto fill maxPriorityFeePerGas and maxFeePerGas
     if (options.type === CONST.TRANSACTION_TYPE_EIP1559) {
-      if (!options.maxPriorityFeePerGas || !options.maxFeePerGas) {
-        options.maxPriorityFeePerGas = await this.maxPriorityFeePerGas();
+      if (options.gasPrice) {
+        options.maxFeePerGas = options.gasPrice;
+        options.maxPriorityFeePerGas = options.gasPrice;
+        options.gasPrice = undefined;
+      }
 
+      if (!options.maxPriorityFeePerGas) {
+        options.maxPriorityFeePerGas = await this.maxPriorityFeePerGas();
+      }
+
+      if (!options.maxFeePerGas) {
         if (!baseFeePerGas) {
           const block = await this.getBlockByEpochNumber(options.epochHeight, false);
           baseFeePerGas = block.baseFeePerGas;
         }
 
-        options.maxFeePerGas = options.maxPriorityFeePerGas + (baseFeePerGas * BigInt(4)) / BigInt(3);
+        options.maxFeePerGas = options.maxPriorityFeePerGas + baseFeePerGas * BigInt(2);
+      }
+
+      if (options.maxFeePerGas < options.maxPriorityFeePerGas) {
+        throw new Error('`maxFeePerGas` should not be less than `maxPriorityFeePerGas`');
       }
     }